requirements.txt
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
python-qt-binding==1.1.2
PyYAML==6.0.1
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
typeguard==4.3.0
typing_extensions==4.12.2

